08_final_project/project_template/pi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
paho-mqtt==1.6.1
pydantic==2.5.0
python-dotenv==1.0.0
